Move bulk import logic into bulk creator
